requirements.txt
numpy >= 1.1, < 2.0
scipy >= 0.13, < 0.17
pandas >= 0.13, < 0.18
scikit-learn >= 0.17, < 0.18
nose >= 1.3.1, < 1.4
Flask==0.11
imutils >= 0.3.6
pillow==3.3.0
flask-socketio==2.6
gunicorn==19.6.0
gevent >= 1.1.2
futures==3.0.5
Flask-SQLAlchemy
mysql-python
Flask-Migrate